refactored how repliers are loaded into the dispatcher
