+++ giru/__main__.py
@@ -1,6 +1,5 @@
 import logging
 import os
-from os.path import join
 
 from telegram import ParseMode
 from telegram.ext import CommandHandler, MessageHandler, Filters, Updater
@@ -21,12 +20,13 @@ from giru.commands import (
 )
 from giru.core.repliers import load_repliers_from_csv_file
 from giru.repliers import *
+from giru.saved_reply_storage import InMemoryReplyStorageProvider, FileSystemReplyStorageProvider
 from giru.settings import (
     FIREBASE_ACCOUNT_KEY_FILE_PATH,
     FIREBASE_DATABASE_URL,
     GIRU_STORAGE_LOCATION,
 )
-from giru.settings import TELEGRAM_TOKEN, GIRU_DATA_PATH, REPLIES_FILE_PATH
+from giru.settings import TELEGRAM_TOKEN, REPLIES_FILE_PATH
 
 logging.basicConfig(
     format="%(asctime)s - %(name)s - %(levelname)s - %(message)s", level=logging.INFO
@@ -34,7 +34,7 @@ logging.basicConfig(
 updater = Updater(token=TELEGRAM_TOKEN)
 
 # NOTE: Replies are being saved in new-line delimited JSON (.ndjson)
-message_storage = FileSystemReplyStorageProvider(join(GIRU_DATA_PATH, "replies.ndjson"))
+message_storage = FileSystemReplyStorageProvider(SAVED_REPLIES_FILE_PATH)
 if GIRU_STORAGE_LOCATION == "in_memory":
     message_storage = InMemoryReplyStorageProvider()
 elif GIRU_STORAGE_LOCATION == "firebase":
@@ -84,25 +84,7 @@ message_handlers = [
     MessageHandler(FilterScores(), record_points),
     MessageHandler(FilterRecon(), recon),
     MessageHandler(FilterReplyToGiru(), send_reply_to_user),
-    mmg_replier.to_message_handler(),
-    cposp_replier.to_message_handler(),
-    wtf_replier.to_message_handler(),
-    mentira_replier.to_message_handler(),
-    hbd_replier.to_message_handler(),
-    salute_gay_replier.to_message_handler(),
-    salute_hola_replier.to_message_handler(),
-    salute_klk_replier.to_message_handler(),
-    salute_klk_post_replier.to_message_handler(),
-    diablo_replier.to_message_handler(),
-    calmate_filter.to_message_handler(),
-    felicidades_filter.to_message_handler(),
-    haters_replier.to_message_handler(),
-    ok_gracia_replier.to_message_handler(),
-    todo_bien_replier.to_message_handler(),
-    menor_replier.to_message_handler(),
-    alcohol_replier.to_message_handler(),
-    familia_replier.to_message_handler(),
-    droga_replier.to_message_handler(),
+    *(m.to_message_handler() for m in built_in_repliers)
 ]
 
 for msg_h in message_handlers:
@@ -114,15 +96,13 @@ try:
             dp.add_handler(r.to_message_handler())
 except FileNotFoundError:
     logging.error(
-        '[ERROR] replies file "{}" not found, file-based replies will not be triggered.'.format(
-            REPLIES_FILE_PATH
-        )
+        "replies file %r not found, file-based replies will not be triggered.",
+        REPLIES_FILE_PATH,
     )
 except ValueError:
     logging.error(
-        '[ERROR] replies file "{}" cannot be processed, file-based replies will not be triggered.'.format(
-            REPLIES_FILE_PATH
-        )
+        "replies file %r cannot be processed, file-based replies will not be triggered.",
+        REPLIES_FILE_PATH,
     )
 
 dp.add_handler(MessageHandler(Filters.command, unknown))
